requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.28.2
gunicorn==20.1.0
heroku3==5.1.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
patsy==0.5.2
Pillow==8.4.0
plotly==5.4.0
protobuf==3.17.3
psycopg2-binary==2.9.3
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
statsmodels==0.13.1
tenacity==8.0.1
threadpoolctl==3.0.0
tomli==1.2.2
urllib3==1.26.7
Werkzeug==2.0.2
xgboost==1.5.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.28.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
patsy==0.5.2
Pillow==8.4.0
plotly==5.4.0
protobuf==3.17.3
psycopg2-binary==2.9.3
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
statsmodels==0.13.1
tenacity==8.0.1
threadpoolctl==3.0.0
tomli==1.2.2
urllib3==1.26.7
Werkzeug==2.0.2
xgboost==1.5.0
